refactor(ui): split api barrels into domain modules
Split api/index.js (508L) into projects.js, evaluations.js, scores.js,
gradeFormula.js, providers.js, localModels.js, updates.js. Split
api/shared.js (333L) into sharedStatus.js, sharedProjectData.js,
sharedPublish.js. Both become export barrels so all existing import
paths keep working.

Moved epochSecondsToMs out of shared.js private scope and exported it
from sharedStatus.js so sharedProjectData.js can reuse it without
duplicating the conversion logic.

Dropped api/index.js and api/shared.js from the size grandfather list
(both now well under the 300-line limit) and lowered the grandfather
count-lock and size baseline to match.
