+++ src/quodeq/ui/src/api/shared.js
@@ -2,331 +2,12 @@
  * Shared repository API client — read-only mirrors of the project read endpoints,
  * plus config management (connect, disconnect, refresh, status) and publish/pull.
  *
- * Timestamp units: the backend (services/shared_repo.py's published_meta and
- * last_synced_at) sends publishedAt/lastSynced as UNIX epoch SECONDS (git log
- * `%ct` is an int; st_mtime is a float) -- but every "N ago" consumer
- * (relativeTime in components/LastFetchedLine.jsx) expects milliseconds, same
- * as Date.now()/`new Date(ms)`. Converting seconds->ms is done once, here, at
- * the API-client boundary, so every consumer downstream always sees ms and
- * never has to know the wire units. Passing raw seconds through would render
- * as a 1970 date ("57 years ago") -- see epochSecondsToMs below.
+ * This module is a barrel: the actual implementations live in domain
+ * modules (sharedStatus.js, sharedProjectData.js, sharedPublish.js),
+ * re-exported here for backward compatibility so every existing
+ * `from './shared.js'` import keeps working unchanged.
  */
 
-import { request, BASE } from './request.js';
-import { createProject } from '../models/project.js';
-import { createDashboard } from '../models/dashboard.js';
-import { createDimension, createDimensionEval, createSlimDimension } from '../models/dimension.js';
-
-/**
- * Convert a UNIX epoch-seconds timestamp (as sent by the backend) to
- * epoch-milliseconds (as expected by every "N ago" / relativeTime consumer).
- * Null/absent/0 all normalize to null -- there is no meaningful "N ago" for
- * an unset timestamp, and 0 never occurs as a real value here.
- * @param {number|null|undefined} seconds
- * @returns {number|null}
- */
-function epochSecondsToMs(seconds) {
-  return typeof seconds === 'number' && seconds ? seconds * 1000 : null;
-}
-
-// ── Config Management ───────────────────────────────────────────────────────
-
-/**
- * Get the shared repository connection status.
- * @returns {Promise<{configured: boolean, url: string|null, lastSynced: number|null, publish: Object}>}
- *   lastSynced is epoch-milliseconds (converted from the backend's epoch
- *   seconds; see epochSecondsToMs). `publish.finishedAt`, if present, is
- *   passed through unconverted (raw epoch seconds) -- no UI consumer currently
- *   formats it as a date.
- */
-export async function getSharedStatus() {
-  const data = await request('/shared/status');
-  return {
-    ...data,
-    lastSynced: epochSecondsToMs(data?.lastSynced),
-  };
-}
-
-/**
- * Connect to a shared repository.
- * @param {string} url - Git repository URL
- * @returns {Promise<{configured: boolean, url: string}>}
- */
-export function connectShared(url) {
-  return request('/shared/config', {
-    method: 'PUT',
-    body: JSON.stringify({ url }),
-  });
-}
-
-/**
- * Disconnect from the shared repository.
- * @returns {Promise<{configured: boolean}>}
- */
-export function disconnectShared() {
-  return request('/shared/config', {
-    method: 'DELETE',
-  });
-}
-
-/**
- * Refresh the shared repository (fetch latest changes).
- * @returns {Promise<{stale: boolean, lastSynced: string}>}
- */
-export function refreshShared() {
-  return request('/shared/refresh', {
-    method: 'POST',
-  });
-}
-
-// ── Project List & Info ─────────────────────────────────────────────────────
-
-/**
- * List projects from the shared repository.
- * Unlike listProjects's envelope (projects + warmup), this envelope carries sync metadata
- * because the shared tab needs lastSynced and stale status.
- * @param {{refresh?: boolean}} [options={}]
- * @returns {Promise<{projects: import('../models/project.js').Project[], lastSynced: number|null, stale: boolean}>}
- *   Both the envelope's lastSynced and each project's publishedAt are
- *   epoch-milliseconds (converted from the backend's epoch seconds; see
- *   epochSecondsToMs).
- */
-export async function sharedListProjects({ refresh = false } = {}) {
-  const refreshParam = refresh ? '1' : '0';
-  const data = await request(`/shared/projects?refresh=${refreshParam}`);
-  const list = data?.projects ?? data ?? [];
-  const projects = Array.isArray(list) ? list.map(createProject) : [];
-
-  // Pass through shared-specific metadata
-  if (Array.isArray(list)) {
-    projects.forEach((proj, idx) => {
-      if (list[idx]) {
-        proj.publishedBy = list[idx].publishedBy ?? null;
-        proj.publishedAt = epochSecondsToMs(list[idx].publishedAt);
-        proj.source = list[idx].source ?? 'shared';
-      }
-    });
-  }
-
-  return {
-    projects,
-    lastSynced: epochSecondsToMs(data?.lastSynced),
-    stale: data?.stale ?? false,
-  };
-}
-
-/**
- * Get detailed info for a shared project.
- * @param {string} projectId
- * @returns {Promise<import('../models/project.js').Project>}
- *   Like sharedListProjects, publishedBy/publishedAt are passed through
- *   after createProject() (which only knows the base Project shape and
- *   would otherwise silently drop them), with publishedAt converted from
- *   the backend's epoch seconds to epoch-milliseconds.
- */
-export async function sharedGetProjectInfo(projectId) {
-  const data = await request(`/shared/projects/${encodeURIComponent(projectId)}/info`);
-  const project = createProject(data);
-  project.publishedBy = data?.publishedBy ?? null;
-  project.publishedAt = epochSecondsToMs(data?.publishedAt);
-  project.source = data?.source ?? 'shared';
-  return project;
-}
-
-/**
- * Get runs for a shared project.
- * @param {string} projectId
- * @returns {Promise<{runs: Array}>}
- */
-export function sharedGetRuns(projectId) {
-  return request(`/shared/projects/${encodeURIComponent(projectId)}/runs`);
-}
-
-// ── Dashboard & Scores ──────────────────────────────────────────────────────
-
-/**
- * Get dashboard for a shared project run.
- * @param {string} projectId
- * @param {string} [run='latest']
- * @returns {Promise<import('../models/dashboard.js').Dashboard>}
- */
-export async function sharedGetDashboard(projectId, run = 'latest') {
-  const q = run ? `?run=${encodeURIComponent(run)}` : '';
-  const data = await request(`/shared/projects/${encodeURIComponent(projectId)}/dashboard${q}`);
-  return createDashboard(data);
-}
-
-/**
- * Slim compare-summary for a shared project — the /compare-summary payload
- * served from the shared clone's own evaluations root, shape-identical to
- * the local endpoint so the Compare tab can mix sources row by row.
- * @param {string} projectId
- * @returns {Promise<{project: string, summary: Object, dimensions: Array, trend: Array, runsCount: number, lastRun: Object|null}>}
- */
-export async function sharedGetCompareSummary(projectId) {
-  const data = await request(`/shared/projects/${encodeURIComponent(projectId)}/compare-summary`);
-  if (Array.isArray(data?.dimensions)) data.dimensions = data.dimensions.map(createSlimDimension);
-  return data;
-}
-
-/**
- * Get accumulated scores for a shared project.
- * @param {string} projectId
- * @param {string} [asOfRun=null]
- * @returns {Promise<Object>}
- */
-export async function sharedGetAccumulated(projectId, asOfRun = null) {
-  const q = asOfRun ? `?asOf=${encodeURIComponent(asOfRun)}` : '';
-  const data = await request(`/shared/projects/${encodeURIComponent(projectId)}/accumulated${q}`);
-  if (data && Array.isArray(data.dimensions)) {
-    data.dimensions = data.dimensions.map(createDimension);
-  }
-  return data;
-}
-
-/**
- * Get unified scores for a shared project.
- * @param {string} projectId
- * @param {string} [asOfRun=null]
- * @returns {Promise<{accumulated: Object, trend: Array, availableRuns: Array}>}
- */
-export async function sharedGetProjectScores(projectId, asOfRun = null) {
-  const q = asOfRun ? `?asOf=${encodeURIComponent(asOfRun)}` : '';
-  const data = await request(`/shared/projects/${encodeURIComponent(projectId)}/scores${q}`);
-  if (data?.accumulated && Array.isArray(data.accumulated.dimensions)) {
-    data.accumulated.dimensions = data.accumulated.dimensions.map(createDimension);
-  }
-  return data;
-}
-
-/**
- * Get slim scores for a specific run.
- * @param {string} projectId
- * @param {string} runId
- * @returns {Promise<{dimensions: Array, summary: Object}>}
- */
-export async function sharedGetRunScores(projectId, runId) {
-  const data = await request(
-    `/shared/projects/${encodeURIComponent(projectId)}/scores/${encodeURIComponent(runId)}`
-  );
-  if (Array.isArray(data?.dimensions)) data.dimensions = data.dimensions.map(createSlimDimension);
-  return data;
-}
-
-// ── Dimension Eval & Violations ─────────────────────────────────────────────
-
-/**
- * Get dimension evaluation details for a shared project.
- * @param {string} projectId
- * @param {string} runId
- * @param {string} dimension
- * @returns {Promise<import('../models/dimension.js').DimensionEval>}
- */
-export async function sharedGetDimensionEval(projectId, runId, dimension) {
-  const data = await request(
-    `/shared/projects/${encodeURIComponent(projectId)}/dimensions/${encodeURIComponent(dimension)}/eval?run=${encodeURIComponent(runId)}`
-  );
-  return createDimensionEval(data);
-}
-
-/**
- * Get violations for a shared project run.
- * @param {string} projectId
- * @param {string} runId
- * @returns {Promise<Object>}
- */
-export function sharedGetViolations(projectId, runId) {
-  return request(
-    `/shared/projects/${encodeURIComponent(projectId)}/violations?run=${encodeURIComponent(runId)}`
-  );
-}
-
-// ── Findings (read-only mirrors) ────────────────────────────────────────────
-// Shared projects are read-only in the app — there are no shared mutation
-// routes (dismiss/restore/delete/unverify), only these list mirrors so the
-// dismissed/verified sub-tabs can display a shared project's existing state.
-
-// Mirrors the local listDismissedFindings' server-side hard cap (see
-// api/findings.js DISMISSED_REQUEST_LIMIT) so a shared project's dismissed
-// list isn't silently truncated to the API's default page size.
-const SHARED_DISMISSED_REQUEST_LIMIT = 5000;
-
-/**
- * List dismissed findings for a shared project.
- * @param {string} projectId
- * @returns {Promise<Array>} Dismissed findings array (same item shape as listDismissedFindings)
- */
-export function sharedListDismissedFindings(projectId) {
-  return request(
-    `/shared/projects/${encodeURIComponent(projectId)}/findings/dismissed`
-    + `?limit=${SHARED_DISMISSED_REQUEST_LIMIT}`
-  );
-}
-
-/**
- * List verified-badge entries for a shared project.
- * @param {string} projectId
- * @returns {Promise<Array>} Entries: { req, file, line, note, verifiedAt }
- */
-export function sharedListVerifiedFindings(projectId) {
-  return request(`/shared/projects/${encodeURIComponent(projectId)}/findings/verified`);
-}
-
-// ── Publish & Pull ──────────────────────────────────────────────────────────
-
-/**
- * Publish a local project to the shared repository.
- * @param {string} projectId
- * @returns {Promise<{started: boolean}>}
- */
-export function publishProject(projectId) {
-  return request(`/projects/${encodeURIComponent(projectId)}/publish`, {
-    method: 'POST',
-  });
-}
-
-/**
- * Pull a shared project into the local evaluations.
- *
- * Uses raw fetch (not the shared request() wrapper) so a 409 collision
- * response can be read as err.status/err.kind/err.existingProjectId/
- * err.projectName -- same contract as importProject() in api/index.js,
- * since both funnel through the backend's import_zip_stream. The
- * "pull local copy" footer action on the online Projects tab needs
- * err.status === 409 to show its inline copy-confirm affordance.
- *
- * @param {string} projectId
- * @param {string} [action] - 'copy' or 'replace' (resolves a 409 collision)
- * @returns {Promise<{imported: boolean, projectId: string}>}
- * @throws {Error & { status: number, code?: string, kind?: string, existingProjectId?: string, projectName?: string }} on non-2xx
- */
-export async function pullSharedProject(projectId, action) {
-  const body = action ? { action } : {};
-  let res;
-  try {
-    res = await fetch(`${BASE}/shared/projects/${encodeURIComponent(projectId)}/pull`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-      // Generous: a pull imports a zip stream from the shared repository,
-      // but a stalled connection must not leave the pull pending forever.
-      signal: AbortSignal.timeout(600000), // 10 min
-    });
-  } catch (e) {
-    if (e?.name === 'TimeoutError' || e?.name === 'AbortError') {
-      throw new Error('Pull timed out. Check the shared repository connection and try again.');
-    }
-    throw e;
-  }
-  const payload = await res.json().catch(() => ({}));
-  if (!res.ok) {
-    const err = new Error(payload.error || `pullSharedProject failed (${res.status})`);
-    err.status = res.status;
-    if (payload.code) err.code = payload.code;
-    if (payload.kind) err.kind = payload.kind;
-    if (payload.existingProjectId) err.existingProjectId = payload.existingProjectId;
-    if (payload.projectName) err.projectName = payload.projectName;
-    throw err;
-  }
-  return payload;
-}
+export * from './sharedStatus.js';
+export * from './sharedProjectData.js';
+export * from './sharedPublish.js';
