+++ src/quodeq/ui/tools/size_grandfather.mjs
@@ -10,10 +10,8 @@ export const SIZE_GRANDFATHER = [
   "src/App.test.jsx",
   "src/api/applyMutationDelta.js",
   "src/api/applyMutationDelta.test.jsx",
-  "src/api/index.js",
   "src/api/index.test.jsx",
   "src/api/queryKeys.test.jsx",
-  "src/api/shared.js",
   "src/api/shared.test.jsx",
   "src/components/ContextBlock.jsx",
   "src/components/HeatGridCells.jsx",
